api/requirements.txt
alembic==0.9.6
aniso8601==8.0.0
appdirs==1.4.4
bcrypt==3.1.4
Cerberus==1.3.2
cffi==1.14.2
click==6.7
coverage==4.4.2
distlib==0.3.1
enum-compat==0.0.3
eventlet==0.21.0
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.1.1
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
greenlet==0.4.16
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.11.2
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-jsonapi==0.23.2
psycopg2-binary==2.8.5
pycparser==2.18
PyJWT==1.7.1
python-dateutil==2.6.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2017.3
selenium==3.8.1
six==1.11.0
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
uWSGI==2.0.19.1
virtualenv==20.0.31
Werkzeug==1.0.1